% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_observed_nights}
\alias{add_observed_nights}
\title{Add Observed Night using recording_time}
\usage{
add_observed_nights(df)
}
\arguments{
\item{df}{Dataframe create from either get_sa_bulk_wavs() or
get_ma_bulk_wavs()}
}
\description{
Adds an observed_night field to the dataframe using the recording_time
field.  Uses Noon as the cut off for previous vs current night
}
\examples{

\dontrun{
# Be sure to run df = clean_time_fields(df) first
df = add_observed_nights(df)
}

}
