% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_start_end_nights}
\alias{add_start_end_nights}
\title{Add Start and End nights for an Acoustic Survey}
\usage{
add_start_end_nights(df)
}
\arguments{
\item{df}{Dataframe create from either get_sa_bulk_wavs() or
get_ma_bulk_wavs()}
}
\description{
Uses survey_start_time and survey_end_time to create 2 new fields
(survey_night_start and survey_night_end) from cleaned acoustic data.
Be sure to run clean_time_frames() on the input dataframe before
running this function.
}
\examples{

\dontrun{
# Be sure to run df = clean_time_fields(df) first
df = add_start_end_nights(df)
}

}
