% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\docType{data}
\name{bats_df}
\alias{bats_df}
\title{NABat species lookup table}
\format{An object of class \code{data.frame} with 83 rows and 7 columns.}
\usage{
bats_df
}
\description{
Reads in dataframe for NABat species lookup table
}
\examples{

nabatr::bats_df
}
\keyword{NABat}
\keyword{bats,}
\keyword{species,}
