% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_ac_doc}
\alias{build_ac_doc}
\title{Build Report document in .docx file for Acoustic Stationary Project Data}
\usage{
build_ac_doc(out_dir_, file_name_,
  title_ = "Implementing NABat on National Wildlife Refuges in the Pacific Northwest Region (R1): Results from 2016 - 2018",
  by_ = "Jenny Barnett",
  member_role_org_ = "Zone Inventory and Monitoring Biologist",
  date_ = format(Sys.time(), "\%B \%d, \%Y"), map = NULL)
}
\description{
Build Report document in .docx file for Acoustic Stationary Project Data
}
