% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{build_cc_table_1}
\alias{build_cc_table_1}
\title{Build Colony Count Table 1}
\usage{
build_cc_table_1(cc_bulk_df, noid = TRUE)
}
\arguments{
\item{noid}{Boolean Whether or not to include NoID in table}

\item{cc_bulkd_df}{Dataframe colony count bulk data}
}
\description{
Summary of winter colony count surveys
}
