% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{build_cc_table_2}
\alias{build_cc_table_2}
\title{Build Colony count table 2}
\usage{
build_cc_table_2(cc_bulk_df)
}
\arguments{
\item{cc_bulkd_df}{Dataframe colony count bulk data}
}
\description{
Number of species found dead
and alive at each GRTS cell
}
