% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_request_queries.R
\name{build_data_request}
\alias{build_data_request}
\title{Submit Build Data Request}
\usage{
build_data_request(
  token,
  data_request_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{data_request_id}{Integer that specifies data request identifier}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
Submits a build data request to Nabat that adds a new Zip file
for this data request id.
}
