% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_grts_plot}
\alias{build_grts_plot}
\title{Build Plot for GRTS or Sites Acoustic Stationary Count data}
\usage{
build_grts_plot(ordered_combined_data, type = "sites")
}
\description{
uses the ordered_combined_data that contains bat counts for
species with values >0 at that location and then creates a plot to
visualize
}
