% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_ma_doc}
\alias{build_ma_doc}
\title{Build Mobile Acoustic Report document in .docx file}
\usage{
build_ma_doc(
  out_dir,
  file_name = "ma_doc.docx",
  project_df,
  project_id,
  ma_bulk_df,
  species_df,
  year,
  nightly_observed_list,
  date = format(Sys.time(), "\%B \%d, \%Y")
)
}
\description{
Using the outputs from get_projects(), get_ma_project_summary(), get_ma_bulk_wavs(),
and get_observed_nights() this function will create a report .docx file to an out_dir.
}
