% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{build_ma_figure_1}
\alias{build_ma_figure_1}
\title{Build Mobile Acoustic Report Figure 1}
\usage{
build_ma_figure_1(ma_bulk_df, project_id, project_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{project_id}{Integer project id}

\item{project_df}{Dataframe project_df}

\item{year}{Integer (optional) Project year with mobile acoustic data}
}
\description{
Creates a leaflet Map for mobile acoustic data
}
\examples{

\dontrun{
ma_figure_1 = build_ma_figure_1(ma_bulk_df, year)
}

}
