% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_report_helpers.R
\name{build_ma_table_1}
\alias{build_ma_table_1}
\title{Build Mobile Acoustic Report Table 1}
\usage{
build_ma_table_1(ma_bulk_df, project_id, project_df, species_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe Mobile acoustic bulk data. Can get from
get_ma_bulk_wavs() in nabatr}

\item{project_id}{Integer project id}

\item{project_df}{Dataframe project_df}

\item{species_df}{Dataframe species nabat lookup dataframe}

\item{year}{Integer Project year with mobile acoustic data}
}
\description{
Returns a Table and its description. Table
with NABat GRTS cells surveyed and
Number of unique mobile transect routes, detector nights,
and species detected are also shown for each cell.
}
