% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{build_ma_table_2}
\alias{build_ma_table_2}
\title{Build Mobile Acoustic Report Table 2}
\usage{
build_ma_table_2(ma_bulk_df, species_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{species_df}{Dataframe species dataframe}

\item{year}{Integer (optional) Project year with mobile acoustic data}
}
\description{
Bat species table for this input ma_bulk_df
}
\examples{

\dontrun{
ma_table_2 = build_ma_table_2(ma_bulk_df, species_df, year)
}

}
