% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_report_helpers.R
\name{build_ma_table_3}
\alias{build_ma_table_3}
\title{Build Mobile Acoustic Report Table 3}
\usage{
build_ma_table_3(ma_bulk_df, nightly_observed_list, species_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{nightly_observed_list}{(optional) List from running get_observed_nights()}

\item{species_df}{Dataframe Species dataframe}

\item{year}{Integer (optional) Project year with mobile acoustic data}
}
\description{
Potentially returns 2 tables and their descriptions.
Table 3A is for the auto detected species table and table 3B is
for the manual detected species table.
}
