% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_sa_doc}
\alias{build_sa_doc}
\title{Build Stationary Acoustic Report document in .docx file}
\usage{
build_sa_doc(
  out_dir,
  file_name = "sa_doc.docx",
  project_df,
  project_id,
  sa_bulk_df,
  sa_survey_df,
  species_df,
  selected_year,
  nightly_observed_list = NULL,
  date = format(Sys.time(), "\%B \%d, \%Y"),
  range_maps = TRUE
)
}
\description{
Using the outputs from get_projects(), get_sa_project_summary(), get_as_bulk_wavs(),
and get_observed_nights() this function will create a report .docx file to an out_dir.
}
\examples{

\dontrun{
See vignette: build_sa_report (Build Stationary Acoustic Report)
}

}
