% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_figure_1}
\alias{build_sa_figure_1}
\title{Build stationary acoustic figure 1 for report}
\usage{
build_sa_figure_1(
  sa_bulk_df,
  out_dir,
  project_df,
  project_id,
  survey_df,
  selected_year,
  save_bool = FALSE
)
}
\description{
Returns leaflet map
}
