% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_figure_4}
\alias{build_sa_figure_4}
\title{Build stationary acoustic figure 4 for report}
\usage{
build_sa_figure_4(
  sa_bulk_df,
  out_dir,
  species_df,
  selected_year,
  save_bool = FALSE
)
}
\description{
Returns 1 plot with grts and the number of bat passes they had
}
