% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_table_3}
\alias{build_sa_table_3}
\title{Build stationary acoustic table 3 for report}
\usage{
build_sa_table_3(sa_bulk_df, selected_year, species_df)
}
\arguments{
\item{sa_bulk_df}{Dataframe Mobile acoustic bulk data. Can get from
get_ma_bulk_wavs() in nabatr}

\item{selected_year}{Integer Report year to use}

\item{species_df}{Dataframe species nabat lookup dataframe}
}
\description{
Returns a table with GRTS, Species_Detected,
and Method_of_Species_ID
}
