% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{cc_fig_1_ggplot}
\alias{cc_fig_1_ggplot}
\title{Display ggplot Colony count figure 1}
\usage{
cc_fig_1_ggplot(fig)
}
\description{
Use this to display returned figures in list from
build_cc_figure_1
}
