% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{does_sa_type_exist}
\alias{does_sa_type_exist}
\title{Find if sa_bulk_df has manual or auto data present}
\usage{
does_sa_type_exist(sa_bulk_df, species_df, type)
}
\description{
Returns 1 plot with grts and the number of bat
passes they had
}
