% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simpleCap}
\alias{.simpleCap}
\title{Get Stationary Acoustic Results for Report}
\usage{
.simpleCap(x)
}
\description{
Combine the spatial information (states and counties) with the detector info and species detected
}
