% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_request_queries.R
\name{download_data_request}
\alias{download_data_request}
\title{Download Data Request}
\usage{
download_data_request(
  token,
  key,
  output_file,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{key}{String pulled from field Key in get_data_request_files()}

\item{output_file}{String output directory + zip file and extension,
ex: /path/to/mydirectory/new_file.zip}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
downloads a data request specified using the Key field from
the output of get_data_request_files()
}
