% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_acoustic_bulk_df}
\alias{get_acoustic_bulk_df}
\title{Get Acoustic stationary bulk upload template dataframe for a project}
\usage{
get_acoustic_bulk_df(token, username, survey_df, project_id)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{username}{String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}

\item{survey_df}{Dataframe a survey dataframe from the output of get_project_surveys}

\item{project_id}{Numeric or String a project id}
}
\description{
Returns all surveys within a single project (project_id)
}
\examples{

acoustic_bulk_df = get_acoustic_bulk_df(username   = 'NABat_Username',
                                      token      = 'generated-nabat-gql-token',
                                      survey_df  = 'dataframe from output of get_project_surveys()',
                                      project_id = 'number or string of a number')

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
