% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_queries.R
\name{get_acoustic_project_summary}
\alias{get_acoustic_project_summary}
\title{Get a Project's Stationary Acoustic Summary Data}
\usage{
get_acoustic_project_summary(
  token,
  project_df,
  project_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{project_df}{Dataframe project dataframe for a nabat project}

\item{project_id}{Numeric project id for a nabat project}

\item{branch}{(optional) String 'prod' | 'dev' | 'beta' | 'local'}

\item{url}{(optional) String testing option}

\item{aws_gql}{(optional) String option for AWS}

\item{aws_alb}{(optional) String option for AWS}

\item{docker}{(optional) Boolean option for docker}
}
\description{
Returns a dataframe with all of a project's acoustic stationary data down to the event
}
\examples{

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
