% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_acoustic_stationary_report}
\alias{get_acoustic_stationary_report}
\title{Build Report .html file for Acoustic Stationary Project Data}
\usage{
get_acoustic_stationary_report(token, username, project_id, output_dir,
  project_name = NULL, project_description = NULL,
  file_name = "report.html", nightly_plots_type = "grts",
  survey_df = NULL, acoustic_bulk_df = NULL,
  nightly_observed_list = NULL)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{username}{String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}

\item{project_id}{String or Integer project id from NABat ex: 105 or '105'}

\item{output_dir}{String output directory to save report .html file ex: /path/to/directory}

\item{file_name}{(optional) String output file name ex: 'my_report.html' ex: my_report.html}

\item{survey_df}{(optional) Dataframe from running get_project_surveys()}

\item{acoustic_bulk_df}{(optional) Dataframe from running get_acoustic_bulk_wavs()}

\item{manual_nights_df}{(optional) Dataframe from running get_observed_nights()}

\item{auto_nights_df}{(optional) Dataframe from running get_observed_nights()}
}
\description{
Using the outputs from get_projects(), get_project_surveys(), get_acoustic_bulk_wavs(),
and get_observed_nights() this function will create a report .html file to an output_dir.
This function can run with these outputs or without.
}
\examples{

\dontrun{
get_acoustic_stationary_report(token      = 'generated-nabat-gql-token',
                               username   = 'NABat_Username',
                               output_dir = '/path/to/your/output/directory',
                               file_name  = 'report.html',
                               project_id = 'number or string of a number')
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
