% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_queries.R
\name{get_all_project_types}
\alias{get_all_project_types}
\title{Get All project's data types available}
\usage{
get_all_project_types(
  token,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{branch}{(optional) String that defaults to 'prod' but can also be
'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns a dataframe with all of a project's data types
(stationary/mobile/colonycount)
}
