% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{get_all_species_counts_long}
\alias{get_all_species_counts_long}
\title{Build a Long Species count Dataframe at GRTS level for Manual and Automatic}
\usage{
get_all_species_counts_long(auto_nights_df, manual_nights_df, fil = TRUE)
}
\description{
Extracting unique species found for manual or automatic nights acoustic data

\dontrun{

# Both Manual and Auto
all_species_totals_l_l_ = get_all_species_counts_long(auto_nights_df_, manual_nights_df_, fil = TRUE)


}
}
