% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{get_cc_results}
\alias{get_cc_results}
\title{Get Colony Count Results}
\usage{
get_cc_results(cc_bulk_df)
}
\arguments{
\item{cc_bulkd_df}{Dataframe colony count bulk data}
}
\description{
Returns compiled results for this project's Colony Count data
}
