% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{get_cc_species}
\alias{get_cc_species}
\title{Get Colony Count Species info}
\usage{
get_cc_species(cc_bulk_df, species_df, format = "df")
}
\description{
Extracts Get all of the species found within these
Colony Count data
format = 'df' | 'vector' | 'vectorNoId' (remove NoID from list)
}
