% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_colony_bulk_counts}
\alias{get_colony_bulk_counts}
\title{Get Winter colony count bulk upload template dataframe for a project}
\usage{
get_colony_bulk_counts(
  token,
  survey_df,
  project_id,
  branch = "prod",
  url = NULL
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{survey_df}{Dataframe a survey dataframe from the output of get_project_surveys}

\item{project_id}{Numeric or String a project id}
}
\description{
Returns all surveys within a single project (project_id)
}
\examples{

\dontrun{
acoustic_bulk_df = get_acoustic_bulk_wavs(token      = 'generated-nabat-gql-token',
                                        survey_df  = 'dataframe from output of get_project_surveys()',
                                        project_id = 'number or string of a number')
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
