% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_request_queries.R
\name{get_data_request_approvals}
\alias{get_data_request_approvals}
\title{Get All Approvals for this Data Request}
\usage{
get_data_request_approvals(
  token,
  data_request_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{data_request_id}{Integer that specifies data request identifier}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
Returns a dataframe of all of the approvals within this
data request
}
