% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_emergence_bulk_counts}
\alias{get_emergence_bulk_counts}
\title{Get Emergence Count Project Data}
\usage{
get_emergence_bulk_counts(
  token,
  survey_df,
  project_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE,
  return_t = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{survey_df}{Dataframe a survey dataframe from the output
of get_ma_project_summary()}

\item{project_id}{Numeric or String a project id}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}

\item{return_t}{(optional) Boolean Changes the Returned value to a list
and adds a token as one of the returned values (if set to TRUE)}
}
\description{
Returns a dataframe matching the upload bulk template
for Emergence Count data
}
