% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_grts_covariate}
\alias{get_grts_covariate}
\title{Get Covariate data for all GRTS in a Sample Frame}
\usage{
get_grts_covariate(
  token,
  covariate_id,
  sample_frame_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{covariate_id}{Numeric Id for the covariate to query.
This value can be extrapolated from the output of get_covariates()}

\item{sample_frame_id}{Numeric Id for the sample frame to query.
This value can be extrapolated from the output of get_sample_frames()}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
Using a covariate and sample frame id, return a dataframe of
each grts and the information coresponding to that covariate.
}
