% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nabat_grts.R
\name{get_grts_data}
\alias{get_grts_data}
\title{Return the spatial layer from the associated wfs query}
\usage{
get_grts_data(grid_frame, query)
}
\arguments{
\item{grid_frame}{String name of the grid frame to return}

\item{query}{String query to apply to the request ("state_n_1='Florida'")
for query syntac help see: https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}
}
\description{
Return the spatial layer from the associated wfs query
}
