% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_grts_frame_name}
\alias{get_grts_frame_name}
\title{Get Acoustic stationary bulk upload template dataframe for a project}
\usage{
get_grts_frame_name(project_df, project_id)
}
\description{
Get Acoustic stationary bulk upload template dataframe for a project
}
