% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{get_grts_shp}
\alias{get_grts_shp}
\title{Build shapefile from GRTS IDs}
\usage{
get_grts_shp(grts_ids)
}
\arguments{
\item{grts_ids}{Vector GRTS Ids}
}
\description{
Builds a grts shapefile from the grts_ids parameter.  note: uses rgdal and spatial packages.
}
\examples{

\dontrun{
}

}
\keyword{CONUS}
\keyword{NABat,}
\keyword{bats,}
\keyword{grts,}
\keyword{species,}
