% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_ma_batch}
\alias{get_ma_batch}
\title{Get Mobile Acoustic Batch}
\usage{
get_ma_batch(
  token,
  survey_df,
  year = NULL,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{survey_df}{Dataframe MA summary data for a project.  Running
get_ma_project_summary() will create this survey_df}

\item{year}{(optional) String Year of data to extract out.  Can also be set to 'all'
Defaults to NULL which grabs first year available from survey_df.}

\item{branch}{(optional) String that defaults to 'prod' but can also be
'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns all MA batch files at selected year from information
extracted using the survey_df (output of get_ma_project_summary())
}
