% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_ma_bulk_wavs}
\alias{get_ma_bulk_wavs}
\title{Get Acoustic Mobile bulk upload template dataframe
for a project}
\usage{
get_ma_bulk_wavs(
  token,
  survey_df,
  year = NULL,
  batch = "first",
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE,
  return_t = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token()
or get_refresh_token()}

\item{survey_df}{Dataframe a survey dataframe from the output
of get_ma_project_summary()}

\item{year}{(optional) Numeric year of data to be returned.
NULL = first year, 'all' = all years, 2018 = only 2018 data}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}

\item{return_t}{(optional) Boolean Changes the Returned value to a list
and adds a token as one of the returned values (if set to TRUE)}

\item{project_id}{Numeric or String a project id}
}
\description{
Returns all MA surveys within a single project (project_id)
that are parsed by year using survey_df
}
