% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{get_ma_examples}
\alias{get_ma_examples}
\title{Get Mobile Acoustic Report examples}
\usage{
get_ma_examples()
}
\description{
Returns al ist of some preset mobile acoustic report examples
}
\examples{

\dontrun{
ma_report_examples = get_ma_examples()
}

}
