% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nabat_banding_by_states}
\alias{get_nabat_banding_by_states}
\title{Get Bat banding data for States}
\usage{
get_nabat_banding_by_states(
  token,
  states,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{project_id}{Numeric or String a project id}
}
\description{
Returns a dataframe of all the bat banding data from that/those states
}
\examples{

\dontrun{
survey_df = get_project_surveys(token      = 'generated-nabat-gql-token',
                                project_id = 'number or string of a number')
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
