% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nabat_gql_token}
\alias{get_nabat_gql_token}
\title{NABat login to NABAt Database GQL}
\usage{
get_nabat_gql_token(username, password = NULL, branch = "prod", url = NULL)
}
\arguments{
\item{username}{String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}

\item{password}{(optional) String it will prompt you for your password}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}
}
\description{
Get a NABat GQL token to use for queries
}
\examples{

\dontrun{
nabat_gql_token = get_nabat_gql_token(username = 'NABat_Username')
-- Prompts for password
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
