% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nightly_data}
\alias{get_nightly_data}
\title{Get Nightly data for a specific Survey Type}
\usage{
get_nightly_data(
  token,
  sp_code,
  survey_type = "bulk_sae",
  sample_frame = NULL,
  project_id = NULL,
  years = NULL,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{sp_code}{Character NABat species code/s from get_species()}

\item{survey_type}{(optional) String 'bulk_sae' | 'bulk_mae' | 'bulk_cc_hib' | 'bulk_cc_mat'}

\item{sample_frame}{(optional) String 'CONUS' | 'Alaska' | 'Canada' | 'Mexico' |
'Pureto Rico' | 'Hawaii'}

\item{project_id}{Numeric for a project id or multiple project ids}

\item{years}{(optional) Numeric for all years to be extracted for this data}

\item{branch}{(optional) String that defaults to 'prod' but can also be
'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns a dataframe with Detection history (long format)
that can be used to create a wide format per species
}
