% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_night_data.R
\name{get_observed_nights}
\alias{get_observed_nights}
\title{Observed nightly observations for NABat stationary acoustic data}
\usage{
get_observed_nights(acoustic_bulk_df)
}
\arguments{
\item{acoustic_bulk_df}{String it will prompt you for your password}
}
\description{
Creates
}
\examples{

get_observed_nights
}
\keyword{NABat}
\keyword{bats,}
\keyword{species,}
