% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_presigned_data}
\alias{get_presigned_data}
\title{Get presigned data}
\usage{
get_presigned_data(
  project_id,
  token,
  branch = "beta",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or get_refresh_token()}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}

\item{file_path}{String full path to CSV file for preview}

\item{survey_type}{(optional) String 'bulk_sae' | 'bulk_mae' | 'bulk_cc'}
}
\description{
Returns a uuid and presigned url to upload a csv into the AWS bucket
}
