% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project_file_bucket}
\alias{get_project_file_bucket}
\title{Get Project file bucket for AWS}
\usage{
get_project_file_bucket(branch = "prod")
}
\description{
Combine the spatial information (states and counties) with the detector info and species detected
}
