% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_refresh_token}
\alias{get_refresh_token}
\title{NABat login to NABAt Database GQL and get access_token}
\usage{
get_refresh_token(token, branch = "prod", url = NULL, force = FALSE)
}
\arguments{
\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{refresh_token}{String token received in get_nabat_gql_token}
}
\description{
Get a NABat GQL token to use for queries
}
\examples{


}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
