% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_sa_bulk_wavs}
\alias{get_sa_bulk_wavs}
\title{Get Acoustic stationary bulk upload template dataframe
for a project}
\usage{
get_sa_bulk_wavs(
  token,
  survey_df,
  year = NULL,
  batch = "first",
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token()
or get_refresh_token()}

\item{survey_df}{Dataframe a survey dataframe from the output
of get_sa_project_summary()}

\item{year}{(optional) Numeric year of data to be returned.
NULL = first year, 'all' = all years, 2018 = only 2018 data}

\item{batch}{String defaults to 'first' but anything else = 'all'}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
Returns all surveys within a single project (project_id)
that are parsed by year using survey_df
}
