% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_examples}
\alias{get_sa_examples}
\title{Get Stationary Acoustic Report examples}
\usage{
get_sa_examples()
}
\description{
Returns al ist of some preset stationary acoustic
report examples
}
