% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_range_maps}
\alias{get_sa_range_maps}
\title{Get Stationary Acoustic Report Range Maps}
\usage{
get_sa_range_maps(
  sa_bulk_df,
  project_df,
  project_id,
  all_species_totals_l_l,
  species_df,
  out_dir,
  save_bool = TRUE,
  species_shp = species_range_shps
)
}
\arguments{
\item{sa_bulk_df}{Dataframe Mobile acoustic bulk data. Can get from
get_ma_bulk_wavs() in nabatr}

\item{project_df}{Dataframe NABat projects. Can get from
get_projects()}

\item{project_id}{Integer Project Id}

\item{all_species_totals_l_l}{Dataframe Species totals. Can get from
get_all_species_counts_long()}

\item{species_df}{Dataframe species nabat lookup dataframe}

\item{out_dir}{String Directory to save the temp directory with
all the maps into. Be sure not to end with '/'}

\item{save_bool}{Boolean Whether to save out maps or not}

\item{species_shp}{SpatialPolygonsDataFrame Species range maps
for NABat. Auto set to global species_range_shps variable}
}
\description{
Builds and saves out the species grts map with species range
map overlayed and species range map on its own to see full species range.
Using a USGS shapefile for species ranges.
}
