% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_species}
\alias{get_sa_species}
\title{Get Stationary Acoustic Species data}
\usage{
get_sa_species(sa_bulk_df, species_df, type = "all", format = "df")
}
\arguments{
\item{sa_bulk_df}{Dataframe Output from get_sa_

\dontrun{

proj_species_df = get_sa_species(sa_bulk_df, species_df, 'all','df')
proj_species = get_sa_species(sa_bulk_df, species_df, 'all','vectorNoId')
man_proj_species  = get_sa_species(sa_bulk_df, species_df, 'man','vectorNoId')
auto_proj_species = get_sa_species(sa_bulk_df, species_df, 'auto','vectorNoId')

}}
}
\description{
Extracts all of the species found within these
stationary acoustic wav files (for, manual, automatic, or both detection types)
type = 'all' | 'auto' | 'man'
format = 'df' | 'vector' | 'vectorNoId' (remove NoID from list)
}
