% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{get_spdf_from_polys_df}
\alias{get_spdf_from_polys_df}
\title{Build spatial polygons dataframe from GRTS shape dataframe}
\usage{
get_spdf_from_polys_df(grts_shp_df)
}
\arguments{
\item{grts_shp_df}{Dataframe}
}
\description{
Build spatial polygons dataframe from GRTS shape dataframe
}
