% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_species}
\alias{get_species}
\title{Get the bat species lookup table}
\usage{
get_species(
  token,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}
}
\description{
Reads in dataframe for NABat species lookup table
}
\examples{

}
\keyword{NABat}
\keyword{bats,}
\keyword{species,}
