% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{get_species_counts_long}
\alias{get_species_counts_long}
\title{Build a Long Species count Dataframe at GRTS level}
\usage{
get_species_counts_long(nights_df, filter = FALSE)
}
\arguments{
\item{nights_df}{Dataframe acoustic bulk dataframe from output of get_observed_nights().}

\item{filter}{Boolean filters out species that don't have any counts for that species


\dontrun{

# Individual Manual or Auto
manual_species_totals_l_ = get_species_counts_long(manual_nights_df_, filter=TRUE)
auto_species_totals_l_   = get_species_counts_long(auto_nights_df_, filter=TRUE)

}}
}
\description{
Extracting unique species found for manual or automatic nights acoustic data
}
