% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{get_species_counts_wide}
\alias{get_species_counts_wide}
\title{Build a Species count Dataframe}
\usage{
get_species_counts_wide(nightly_df)
}
\arguments{
\item{acoustic_bulk_df}{Dataframe acoustic bulk dataframe from output of get_observed_nights()}
}
\description{
Uses one of the outputs from get_observed_nights() to calculate totals for
number of species present at a GRTS cell or a site ID and then returns a dataframe.
}
\examples{

\dontrun{
species_totals    = get_species_counts(manual_nights_df)
}

}
\keyword{NABat}
\keyword{bats,}
\keyword{species,}
