% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_token_headers}
\alias{get_token_headers}
\title{Get Token Headers}
\usage{
get_token_headers(
  token,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\description{
Create and return correct headers for querying NABat
API. This function also refreshes your NABat GQL token
}
