% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nabat_grts.R
\name{get_wfs_layer_name}
\alias{get_wfs_layer_name}
\title{@description Return the wfs layer name from the WFS associated with a SB item}
\usage{
get_wfs_layer_name(hash)
}
\arguments{
\item{hash}{String SB hash of the item containing the wfs}
}
\description{
@description Return the wfs layer name from the WFS associated with a SB item
}
