% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{grts_coords}
\alias{grts_coords}
\title{Read in GRTS lookup csv for CONUS}
\format{An object of class \code{data.frame} with 81935 rows and 7 columns.}
\usage{
grts_coords
}
\description{
Reads in dataframe that contains all of the GRTS cells and their
4 corners to build shapefiles from.  Projection: WGS84 and extent: CONUS
}
\examples{

nabatr::grts_coords
}
\keyword{CONUS}
\keyword{NABat,}
\keyword{bats,}
\keyword{grts,}
\keyword{species,}
