% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\docType{data}
\name{grts_lookup_df}
\alias{grts_lookup_df}
\title{NABat GRTS lookup list with csvs of coordinates for all GRTS in a region}
\format{An object of class \code{list} of length 6.}
\usage{
grts_lookup_df
}
\description{
Used to grab correct coordinates for GRTS lookups
}
\examples{

\dontrun{
nabatr::grts_lookup_df
}

}
\keyword{GRTS,}
\keyword{NABat}
\keyword{spatial,}
