% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_col}
\alias{move_col}
\title{Move Column in dataframe to other Column location}
\usage{
move_col(data, cols, ref, side = c("before", "after"))
}
\description{
Moves one column with field name x to the location of different
field name y
}
