% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nabat_grts.R
\name{normalize_grid_frame}
\alias{normalize_grid_frame}
\title{normalize the passed grid_frame name to allow for flexibility in calling}
\usage{
normalize_grid_frame(grid_frame)
}
\arguments{
\item{grid_frame}{String name of the grid frame to return}
}
\description{
normalize the passed grid_frame name to allow for flexibility in calling
}
