% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{process_uploaded_csv}
\alias{process_uploaded_csv}
\title{Process uploaded CSV}
\usage{
process_uploaded_csv(
  user_id,
  project_id,
  asUUid,
  template,
  file_name,
  token,
  survey_type = "bulk_sae",
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\description{
Upload a csv to a presigned url
}
