% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{project_dir}
\alias{project_dir}
\title{current directory for reports.R}
\format{An object of class \code{character} of length 1.}
\usage{
project_dir
}
\description{
current directory to use for finding files within package
}
\examples{

}
\keyword{dir,}
\keyword{directory}
