% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_acoustic_df}
\alias{rename_acoustic_df}
\title{Return the acoustic dataframe renamed with appropriate field headers}
\usage{
rename_acoustic_df(acoustic_df)
}
\arguments{
\item{acoustic_df}{Dataframe output from get_acoustic_bulk_df()}
}
\description{
Return the acoustic dataframe renamed with appropriate field headers
}
