% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_colony_df}
\alias{rename_colony_df}
\title{Return the colony dataframe renamed with appropriate field headers}
\usage{
rename_colony_df(colony_df)
}
\arguments{
\item{colony_df}{Dataframe output from get_colony_bulk_df()}
}
\description{
Return the colony dataframe renamed with appropriate field headers
}
