% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_geoms_df}
\alias{rename_geoms_df}
\title{Return the geometry dataframe renamed with appropriate field headers}
\usage{
rename_geoms_df(geoms_df)
}
\arguments{
\item{geoms_df}{dataframe from geom gql query}
}
\description{
Return the geometry dataframe renamed with appropriate field headers
}
