% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_project_df}
\alias{rename_project_df}
\title{Return the project dataframe renamed with appropriate field headers}
\usage{
rename_project_df(project_df)
}
\arguments{
\item{project_df}{Dataframe output from get_projects()}
}
\description{
Return the project dataframe renamed with appropriate field headers
}
