% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_species_df}
\alias{rename_species_df}
\title{Return the species dataframe renamed with appropriate field headers}
\usage{
rename_species_df(species_df)
}
\arguments{
\item{project_df}{Dataframe output from get_species()}
}
\description{
Return the species dataframe renamed with appropriate field headers
}
