% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_survey_df}
\alias{rename_survey_df}
\title{Return the survey dataframe renamed with appropriate field headers}
\usage{
rename_survey_df(survey_df)
}
\arguments{
\item{survey_df}{Dataframe output from get_project_surveys()}
}
\description{
Return the survey dataframe renamed with appropriate field headers
}
