% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\docType{data}
\name{species_range_shps}
\alias{species_range_shps}
\title{NABat species range maps}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 49 rows and 4 columns.
}
\usage{
species_range_shps
}
\description{
Reading in species range maps with readOGR
}
\keyword{datasets}
