% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{upload_csv}
\alias{upload_csv}
\title{Upload CSV to aws}
\usage{
upload_csv(presigned_url, file_path)
}
\arguments{
\item{presigned_url}{String Output from get_presigned_data()}

\item{file_path}{String File path of csv to upload to NABat}
}
\description{
Upload a csv to a presigned url
}
