% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{upload_data}
\alias{upload_data}
\title{Upload CSV Data to NABat Database}
\usage{
upload_data(
  token,
  user_id,
  file_path,
  project_id,
  survey_type,
  file_name = NULL,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{user_id}{Numeric NABat user Id - from get_user_id_by_email()}

\item{file_path}{String full path to CSV file for preview}

\item{project_id}{Numeric or String a project id}

\item{survey_type}{(optional) String 'bulk_sae' | 'bulk_mae' | 'bulk_hib' | 'bulk_mat'}

\item{branch}{(optional) String that defaults to 'prod' but
can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container
or not}
}
\description{
Gets a preview file, presigned URL, UUID, uploads the CSV
to an S3 bucket, and then processes the CSV into the NABat
Database.
}
