% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_col_doc}
\alias{build_col_doc}
\title{Build Report document in .docx file for Colony Count Data}
\usage{
build_col_doc(
  out_dir,
  file_name,
  project_df,
  project_id,
  cc_bulk_df,
  cc_survey_df,
  date = format(Sys.time(), "\%B \%d, \%Y")
)
}
\description{
Using the outputs from get_projects(), get_project_surveys(), get_colony_bulk_counts(),
and () this function will create a report .docx file to an out_dir.
}
