% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_ma_doc}
\alias{build_ma_doc}
\title{Build Mobile Acoustic Report document in .docx file}
\usage{
build_ma_doc(
  out_dir,
  file_name,
  project_df,
  project_id,
  ma_bulk_df,
  species_df,
  year,
  nightly_observed_list,
  date = format(Sys.time(), "\%B \%d, \%Y")
)
}
\arguments{
\item{out_dir}{String output directory to save report .html file ex: /path/to/directory}

\item{file_name}{String output file name ex: paste0('doc_report_',project_id_,'_',Sys.Date(),'.docx')}

\item{project_df}{Dataframe from running get_projects()}

\item{project_id}{Integer project id from NABat}

\item{date}{Date current time in a month/day/Year format ex: format(Sys.time(), "%B %d, %Y")

\dontrun{
doc_ = build_ac_doc(out_dir = '/path/to/output/dir',
                    file_name  = paste0('doc_report_',project_id_,'_',Sys.Date(),'.docx'),
                    project_df = project_df_,
                    project_id = project_id_,
                    ma_bulk_df = ma_bulk_df_,
                    auto_species_grts_df_w,
                    auto_nights_df = auto_nights_df_,
                    manual_nights_df = manual_nights_df_,
                    cover_photo = '/path/to/a/cover/photo.png',
                    map = grts_map,
                    manual_species_grts_df_w = manual_species_grts_df_w_,
                    auto_species_grts_df_w = auto_species_grts_df_w_,
                    auto_species_totals_l = auto_species_totals_l_,
                    manual_species_totals_l = manual_species_totals_l_,
                    date = format(Sys.time(), "%B %d, %Y"),
                    acoustic_bulk_df = acoustic_bulk_df_)
}}

\item{manual_species_grts_df_w}{Dataframe manual species df wide get_species_counts_wide()}

\item{auto_species_grts_df_w}{Dataframe auto species df wide get_species_counts_wide()}
}
\description{
Using the outputs from get_projects(), get_ma_project_summary(), get_ma_bulk_wavs(),
and get_observed_nights() this function will create a report .docx file to an out_dir.
}
