% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{build_ma_figure_3}
\alias{build_ma_figure_3}
\title{Build Mobile Acoustic Report Figure 3}
\usage{
build_ma_figure_3(ma_bulk_df, species_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{species_df}{Dataframe species_df}

\item{year}{Integer (optional) Project year with mobile acoustic data}
}
\description{

}
\examples{

\dontrun{
ma_figure_3 = build_ma_figure_3(ma_bulk_df, species_df, year)
}

}
