% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{build_ma_table_1}
\alias{build_ma_table_1}
\title{Build Mobile Acoustic Report Table 1}
\usage{
build_ma_table_1(ma_bulk_df, species_df, year = NULL)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{species_df}{Dataframe species_df}

\item{year}{Integer Project year with mobile acoustic data}
}
\description{
Returns a Table and its description. Table with NABat GRTS cells surveyed and
Number of unique mobile transect routes, detector nights,
and species detected are also shown for each cell.
}
\examples{

\dontrun{
ma_table_1 = build_ma_table_1(ma_bulk_df, species_df, year)
}

}
