% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_figure_2}
\alias{build_sa_figure_2}
\title{Build stationary acoustic figure 2 for report}
\usage{
build_sa_figure_2(
  sa_bulk_df,
  out_dir,
  species_df,
  selected_year,
  sa_table_3,
  auto_species_grts_df_w,
  manual_species_grts_df_w,
  save_bool = FALSE
)
}
\description{
Returns 2 plots and 2 descriptions
}
