% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_table_1}
\alias{build_sa_table_1}
\title{Build stationary acoustic table 1 for report}
\usage{
build_sa_table_1(sa_bulk_df, project_id, project_df, species_df, selected_year)
}
\description{
Returns a table with GRTS, Species_Detected, and Method_of_Species_ID
}
