% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{build_sa_table_3}
\alias{build_sa_table_3}
\title{Build stationary acoustic table 3 for report}
\usage{
build_sa_table_3(sa_bulk_df, selected_year, species_df)
}
\description{
Returns a table with GRTS, Species_Detected,
and Method_of_Species_ID
}
