% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_acoustic_stationary_report}
\alias{get_acoustic_stationary_report}
\title{Build Report .html file for Acoustic Stationary Project Data}
\usage{
get_acoustic_stationary_report(token, output_dir, project_id,
  file_name = NULL, project_df = NULL, survey_df = NULL,
  acoustic_bulk_df = NULL, manual_nights_df = NULL,
  auto_nights_df = NULL)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{output_dir}{String output directory to save report .html file}

\item{project_id}{String or Integer project id from NABat ex: 105 or '105'}

\item{file_name}{(optional) String output file name ex: 'my_report.html'}

\item{project_df}{(optional) Dataframe from running get_projects()}

\item{survey_df}{(optional) Dataframe from running get_project_surveys()}

\item{acoustic_bulk_df}{(optional) Dataframe from running get_acoustic_bulk_wavs()}

\item{manual_nights_df}{(optional) Dataframe from running get_observed_nights()}

\item{auto_nights_df}{(optional) Dataframe from running get_observed_nights()}
}
\description{
Using the outputs from get_projects(), get_project_surveys(), get_acoustic_bulk_wavs(),
and get_observed_nights() this function will create a report .html file to an output_dir.
This function can run with these outputs or without.
}
\examples{

get_acoustic_stationary_report(username = 'NABat_Username')
-- Prompts for password
}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
