% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report_helpers.R
\name{get_cc_species}
\alias{get_cc_species}
\title{Get Colony Count Species info}
\usage{
get_cc_species(cc_bulk_df, species_df, format = "df")
}
\arguments{
\item{species_df}{Dataframe species nabat lookup dataframe}

\item{format}{String format of data to return 'df' = dataframe of
the species from this data, 'vector' = a string vector of all
of the species from this data, 'vectorNoId' = a strin vector of all
of the species but exclude 'NoID'}

\item{cc_bulkd_df}{Dataframe colony count bulk data}
}
\description{
Extracts Get all of the species found within these
Colony Count data
format = 'df' | 'vector' | 'vectorNoId' (remove NoID from list)
}
