% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_colony_bulk_counts}
\alias{get_colony_bulk_counts}
\title{Get Winter colony count bulk upload template dataframe for a project}
\usage{
get_colony_bulk_counts(
  token,
  survey_df,
  project_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or get_refresh_token()}

\item{survey_df}{Dataframe a survey dataframe from the output of get_acoustic_project_summary()}

\item{project_id}{Numeric or String a project id}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns all surveys within a single project (project_id)
}
\examples{

\dontrun{
acoustic_bulk_df = get_colony_bulk_counts(token,
                                          get_acoustic_project_summary(),
                                          project_id)
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
