% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{get_grts_from_ll}
\alias{get_grts_from_ll}
\title{Get GRTS information from Lat / Lon}
\usage{
get_grts_from_ll(
  token,
  latitude,
  longitude,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{grts_shp_df}{Dataframe}
}
\description{
Takes a latitude and longitude in EPSG 4326 (WGS 84)
and returns a
}
