% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{get_grts_leaflet_map}
\alias{get_grts_leaflet_map}
\title{Build leaflet map for Acoustic Stationary report in CONUS}
\usage{
get_grts_leaflet_map(all_grts, grts_with_data = NULL)
}
\arguments{
\item{all_grts}{Character Vector all grts cell ids found from the survey_df dataframe by running
unique(survey_df$grts_cell_id)}

\item{grts_with_data}{(optional) Character Vector or NULL}
}
\description{
Builds a leaflet map using a vector list of grts cells to add to a leaflet map.  Shows
where the project's grts cells live spatially.  Keep in mind that the project_id
and all_grts must be in CONUS.
}
\examples{

\dontrun{
map = get_grts_leaflet_map(all_grts       = unique(survey_df_$grts_cell_id),
                           grts_with_data = unique(auto_nights_df_$GRTS))
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
