% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{get_grts_shp_df}
\alias{get_grts_shp_df}
\title{Build polygons dataframe from GRTS IDs}
\usage{
get_grts_shp_df(grts_ids)
}
\arguments{
\item{grts_ids}{Character Vector GRTS Ids}
}
\description{
Build polygons dataframe from GRTS IDs
}
