% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{get_ma_results}
\alias{get_ma_results}
\title{Clean Acoustic NABat time fields}
\usage{
get_ma_results(ma_bulk_df, species_df, year)
}
\arguments{
\item{ma_bulk_df}{Dataframe create from either get_ma_bulk_wavs()}

\item{species_df}{Dataframe species_df}

\item{year}{Integer Project year with mobile acoustic data}
}
\description{
Takes a dataframe with these three fieldnames
(recording_time, survey_start_time, and survey_end_time)
from NABat database and converts the date format into a POSIX object
}
\examples{

\dontrun{
get_ma_results(ma_bulk_df, species_df, year)
}

}
