% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nabat_banding_by_states}
\alias{get_nabat_banding_by_states}
\title{Get Bat banding data for States}
\usage{
get_nabat_banding_by_states(
  token,
  states,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or get_refresh_token()}

\item{states}{String Vector state or states.  options:
'Alabama','Alaska','Arizona','Arkansas','California','Colorado','Connecticut','Delaware',
'Florida','Georgia','Hawaii','Idaho','Illinois','Indiana','Iowa','Kansas','Kentucky',
'Louisiana','Maine','Maryland','Massachusetts','Michigan','Minnesota','Mississippi',
'Missouri','Montana','Nebraska','Nevada','New Hampshire','New Jersey','New Mexico',
'New York','North Carolina','North Dakota','Ohio','Oklahoma','Oregon','Pennsylvania',
'Rhode Island','South Carolina','South Dakota','Tennessee','Texas','Utah','Vermont',
'Virginia','Washington','West Virgina','Wisconsin','Wyoming'}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns a dataframe of all the bat banding data from state/states selected
}
\examples{

\dontrun{
survey_df = get_project_surveys(token,
                                project_id)
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
