% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nabat_gql_token}
\alias{get_nabat_gql_token}
\title{NABat login to NABAt Database GQL}
\usage{
get_nabat_gql_token(
  username = NULL,
  password = NULL,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{username}{(optional) String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}

\item{password}{(optional) String it will prompt you for your password}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Get a NABat GQL token to use for queries
}
\examples{

\dontrun{
nabat_gql_token = get_nabat_gql_token(username = 'NABat_Username')
-- Prompts for password
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
