% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_nightly_data}
\alias{get_nightly_data}
\title{Get Nightly data for a specific Survey Type}
\usage{
get_nightly_data(
  token,
  project_id,
  years = "[]",
  survey_type = "bulk_sae",
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{project_id}{Numeric or String a project id}

\item{years}{String or Number This value is a little more confusing since
it gets passed through a JSON and interpreted as an array.  If using a
single year, just do year = 2018.  If using a String, it can be an empty
array or an array with years inside.}

\item{survey_type}{(optional) String 'bulk_sae' | 'bulk_mae' | 'bulk_cc'}

\item{branch}{(optional) String that defaults to 'prod' but can also be
'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Returns a dataframe with Detection history (long format)
that can be used to create a wide format per species
}
