% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_night_data.R
\name{get_observed_nights}
\alias{get_observed_nights}
\title{Observed nightly observations for NABat stationary acoustic data}
\usage{
get_observed_nights(acoustic_bulk_df)
}
\arguments{
\item{acoustic_bulk_df}{Dataframe acoustic bulk dataframe from output of get_acoustic_bulk()}
}
\description{
Creates a list with two dataframes (manual and automatic) nightly count
data at each site id.
}
\examples{

\dontrun{
nightly_observed_list = nabatr::get_observed_nights(acoustic_bulk_df)
manual_nights_df = nightly_observed_list$auto_nightly_df
auto_nights_df   = nightly_observed_list$manual_nightly_df
}

}
\keyword{NABat}
\keyword{bats,}
\keyword{species,}
