% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project_file_bucket}
\alias{get_project_file_bucket}
\title{Get Project file bucket for AWS}
\usage{
get_project_file_bucket(branch = "prod")
}
\arguments{
\item{branch}{String Branch to use either 'beta' or 'prod'}
}
\description{
Returns the name of the AWS bucket based on beta/prod
}
