% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_project_surveys}
\alias{get_project_surveys}
\title{Get a project's Stationary Acoustic Surveys}
\usage{
get_project_surveys(
  token,
  project_df,
  project_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{project_df}{Dataframe output from get_projects()}

\item{project_id}{Numeric or String a project id}
}
\description{
Returns all surveys within a single project (project_id)
}
\examples{

\dontrun{
survey_df = get_project_surveys(token      = 'generated-nabat-gql-token',
                                project_id = 'number or string of a number')
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
