% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_project_surveys}
\alias{get_project_surveys}
\title{Get a project's Accoustic Surveys as Acoustic Bulk Upload format}
\usage{
get_project_surveys(token, username, project_id)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{username}{String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}

\item{project_id}{Numeric or String a project id from}
}
\description{
Returns all surveys within a single project (project_id)
}
\examples{

get_project_surveys = get_projects(username   = 'NABat_Username',
                                   token      = 'generated-nabat-gql-token',
                                   project_id = 'number or string of a number')

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
