% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_project_surveys}
\alias{get_project_surveys}
\title{Get a project's Stationary Acoustic Surveys (DEPRECIATED) see get_acoustic_project_summary()}
\usage{
get_project_surveys(
  token,
  project_df,
  project_id,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or get_refresh_token()}

\item{project_df}{Dataframe output from get_projects()}

\item{project_id}{Numeric or String a project id}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
(DEPRECIATED)
Returns all surveys within a single project (project_id)
}
\examples{

\dontrun{
survey_df = get_project_surveys(token,
                                get_projects(token),
                                project_id)
}

}
\keyword{GQL,}
\keyword{NABat,}
\keyword{Surveys}
\keyword{bats,}
