% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_projects}
\alias{get_projects}
\title{Find your NABat Projects}
\usage{
get_projects(token, username)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}

\item{username}{String your NABat username from https://sciencebase.usgs.gov/nabat/#/home}
}
\description{
Returns all projects that the user has permissions to view
}
\examples{

project_df = get_projects(username = 'NABat_Username',
                          token    = 'generated-nabat-gql-token')

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
