% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_projects}
\alias{get_projects}
\title{Find your NABat Projects}
\usage{
get_projects(
  token,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{token}{String token created from get_nabat_gql_token function}
}
\description{
Returns all projects that the user has permissions to view
}
\examples{

\dontrun{
project_df = get_projects(token = 'generated-nabat-gql-token')
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
