% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{get_refresh_token}
\alias{get_refresh_token}
\title{NABat login to NABAt Database GQL and get access_token}
\usage{
get_refresh_token(
  token,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE,
  force = FALSE
)
}
\arguments{
\item{token}{List token created from get_nabat_gql_token() or get_refresh_token()}

\item{branch}{(optional) String that defaults to 'prod' but can also be 'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Get a NABat GQL token to use for queries
}
\examples{

#' \dontrun{
nabat_gql_token = get_refresh_token(token)
-- Prompts for password
}

}
\keyword{GQL}
\keyword{NABat,}
\keyword{bats,}
