% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_range_maps}
\alias{get_sa_range_maps}
\title{Get Stationary Acoustic Report Range Maps}
\usage{
get_sa_range_maps(
  sa_bulk_df,
  project_df,
  project_id,
  all_species_totals_l_l,
  species_df,
  out_dir,
  save_bool = TRUE
)
}
\description{
builds and saves out the species grts map with species range
map overlayed and species range map on its own to see full species range.
Using a USGS shapefile for species ranges.
}
