% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_results}
\alias{get_sa_results}
\title{Get Stationary Acoustic Report Results}
\usage{
get_sa_results(sa_bulk_df, selected_year, species_df)
}
\arguments{
\item{sa_bulk_df}{Dataframe Mobile acoustic bulk data. Can get from
get_ma_bulk_wavs() in nabatr}

\item{selected_year}{Integer Report year to use}

\item{species_df}{Dataframe species nabat lookup dataframe}
}
\description{
Builds out text with results
}
