% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_report_helpers.R
\name{get_sa_species}
\alias{get_sa_species}
\title{Get Stationary Acoustic Species data}
\usage{
get_sa_species(sa_bulk_df, species_df, type = "all", format = "df")
}
\arguments{
\item{sa_bulk_df}{Dataframe Mobile acoustic bulk data. Can get from
get_ma_bulk_wavs() in nabatr}

\item{species_df}{Dataframe species nabat lookup dataframe}

\item{type}{String type of data to return for Stationary
Acoustic 'auto' | 'man' | 'all' (automatic detection,
manual detections, and both)}

\item{format}{String format of data to return 'df' = dataframe of
the species from this data, 'vector' = a string vector of all
of the species from this data, 'vectorNoId' = a strin vector of all
of the species but exclude 'NoID'

\dontrun{

proj_species_df = get_sa_species(sa_bulk_df, species_df, 'all','df')
proj_species = get_sa_species(sa_bulk_df, species_df, 'all','vectorNoId')
man_proj_species  = get_sa_species(sa_bulk_df, species_df, 'man','vectorNoId')
auto_proj_species = get_sa_species(sa_bulk_df, species_df, 'auto','vectorNoId')

}}
}
\description{
Extracts all of the species found within these
stationary acoustic wav files (for, manual, automatic, or both detection types)
type = 'all' | 'auto' | 'man'
format = 'df' | 'vector' | 'vectorNoId'
(remove NoID + other ambigous bat calls from list)
}
