% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{get_species_counts_long}
\alias{get_species_counts_long}
\title{Build a Long Species count Dataframe at GRTS level}
\usage{
get_species_counts_long(nights_df, filter = FALSE)
}
\description{
Extracting unique species found for manual or automatic nights acoustic data
}
