% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabat_gql_queries.R
\name{process_uploaded_csv}
\alias{process_uploaded_csv}
\title{Process uploaded CSV}
\usage{
process_uploaded_csv(
  user_id,
  project_id,
  asUUid,
  template,
  survey_type,
  token,
  file_path,
  file_name = NULL,
  branch = "prod",
  url = NULL,
  aws_gql = NULL,
  aws_alb = NULL,
  docker = FALSE
)
}
\arguments{
\item{user_id}{Numeric NABat user Id - from get_user_id_by_email()}

\item{project_id}{Numeric or String a project id}

\item{asUUid}{String asUUid - from get_presigned_data()}

\item{template}{Dataframe - from get_upload_file_preview()}

\item{survey_type}{String Type of data to process: 'bulk_sae' | 'bulk_mae'
| 'bulk_hib' | 'bulk_mat' | 'bulk_ee'}

\item{token}{List token created from get_nabat_gql_token() or
get_refresh_token()}

\item{file_path}{String full path to CSV file for preview}

\item{file_name}{String Name of file to be uploaded into NABat website}

\item{branch}{(optional) String that defaults to 'prod' but can also be
'dev'|'beta'|'local'}

\item{url}{(optional) String url to use for GQL}

\item{aws_gql}{(optional) String url to use in aws}

\item{aws_alb}{(optional) String url to use in aws}

\item{docker}{(optional) Boolean if being run in docker container or not}
}
\description{
Process a csv and upload to your NABat Project
}
